requirements.txt
fastapi==0.109.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-multipart==0.0.19
loguru==0.7.2
pydantic==2.5.0
jinja2==3.1.3
aiofiles==23.2.1
requests~=2.32.5